% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_checklist_count.R
\name{map_checklist_count}
\alias{map_checklist_count}
\title{Title}
\usage{
map_checklist_count(x)
}
\arguments{
\item{x}{An sf dataframe with checklist longitudes and latitudes}
}
\value{
A maplibre map of checklist locations
}
\description{
Title
}
