% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Title}
\usage{
mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of values to find the mode for}

\item{na.rm}{A boolean that indicates whether or not to remove missing values}
}
\value{
The modal value of the input vector
}
\description{
Title
}
