% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_breeding_codes.R
\name{summarize_breeding_codes}
\alias{summarize_breeding_codes}
\title{Title}
\usage{
summarize_breeding_codes(x)
}
\arguments{
\item{x}{A dataframe with breeding code data}
}
\value{
A gt() table
}
\description{
Title
}
