% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_effort.R
\name{summarize_effort}
\alias{summarize_effort}
\title{Title}
\usage{
summarize_effort(x)
}
\arguments{
\item{x}{A dataframe of checklist effort time data}
}
\value{
A gt() table
}
\description{
Title
}
