% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_season.R
\name{calc_atlasers}
\alias{calc_atlasers}
\title{Title}
\usage{
calc_atlasers(x)
}
\arguments{
\item{x}{A dataframe with eBird checklists generated via auk}
}
\value{
A dataframe with the distinct count of atlasers in a given block
}
\description{
Title
}
