% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_season.R
\name{calc_breeding_season_coverage}
\alias{calc_breeding_season_coverage}
\title{Title}
\usage{
calc_breeding_season_coverage(x, y)
}
\arguments{
\item{x}{A dataframe with eBird checklists generated via auk}

\item{y}{A dataframe with the PBA3 seasons and calendar month mappings}
}
\description{
Title
}
