% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_season.R
\name{calc_checklist_count}
\alias{calc_checklist_count}
\title{Title}
\usage{
calc_checklist_count(x)
}
\arguments{
\item{x}{A dataframe with eBird checklists generated via auk}
}
\value{
A dataframe with distinct counts of checklists by block
}
\description{
Title
}
