% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_season.R
\name{calc_nocturnal_diurnal_effort}
\alias{calc_nocturnal_diurnal_effort}
\title{Title}
\usage{
calc_nocturnal_diurnal_effort(x, y, z)
}
\arguments{
\item{x}{A dataframe with eBird checklists generated via auk}

\item{y}{A dataframe with the modal observation datetime per checklist}

\item{z}{A dataframe with the sunrise and sunset times for each checklist location}
}
\value{
A dataframe with the diurnal, nocturnal, and unknown effort time.
}
\description{
Title
}
