% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_season.R
\name{calc_nocturnal_species_coded}
\alias{calc_nocturnal_species_coded}
\title{Title}
\usage{
calc_nocturnal_species_coded(x, y)
}
\arguments{
\item{x}{A dataframe with eBird checklists generated via auk.}

\item{y}{A dataframe with the common names of nocturnal species}
}
\description{
Title
}
