% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_season.R
\name{calc_species_coded}
\alias{calc_species_coded}
\title{Title}
\usage{
calc_species_coded(x)
}
\arguments{
\item{x}{A dataframe with eBird checklists generated via auk}
}
\value{
A dataframe with the count of species for each breeding_category_desc in a given block
}
\description{
Title
}
