% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_season.R
\name{calc_species_observed}
\alias{calc_species_observed}
\title{Title}
\usage{
calc_species_observed(x)
}
\arguments{
\item{x}{A dataframe with eBird checklists generated via auk}
}
\value{
A dataframe with the number of distinct species observed in a given block
}
\description{
Title
}
