% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_season.R
\name{summarize_season}
\alias{summarize_season}
\title{Title}
\usage{
summarize_season(
  checklist_df,
  block_df,
  seasons_df = seasons,
  season_filter,
  pba2_block_data = pba2_blocks
)
}
\description{
Title
}
